+++ Box/AudioSampleEntry.h
@@ -11,7 +11,7 @@ class AudioSampleEntry : public Box {
 public:
     AudioSampleEntry(uint32_t size, uint32_t type, uint32_t startPos);
     virtual ~AudioSampleEntry();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
