+++ Box/CompositionTimeToSampleBox.cpp
@@ -4,7 +4,7 @@
 
 #include "CompositionTimeToSampleBox.h"
 
-CompositionTimeToSampleBox::CompositionTimeToSampleBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size,
+CompositionTimeToSampleBox::CompositionTimeToSampleBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size,
                                                                                                                   type,
                                                                                                                   start_pos) {}
 
@@ -12,7 +12,7 @@ CompositionTimeToSampleBox::~CompositionTimeToSampleBox() {
 
 }
 
-void CompositionTimeToSampleBox::Parse(struct Parser *parser, uint32_t start_pos) {
+void CompositionTimeToSampleBox::Parse(struct Parser *parser, uint64_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     FileReader *file_reader = parser->file_reader;
