+++ Box/DefaultContainerBox.cpp
@@ -4,13 +4,13 @@
 
 #include "DefaultContainerBox.h"
 
-DefaultContainerBox::DefaultContainerBox(uint32_t size, uint32_t type, uint32_t start_pos) : Box(size, type, start_pos) {}
+DefaultContainerBox::DefaultContainerBox(uint32_t size, uint32_t type, uint64_t start_pos) : Box(size, type, start_pos) {}
 
 DefaultContainerBox::~DefaultContainerBox() {
 
 }
 
-void DefaultContainerBox::Parse(struct Parser *parser, uint32_t start_pos) {
+void DefaultContainerBox::Parse(struct Parser *parser, uint64_t start_pos) {
     FileReader *file_reader = parser->file_reader;
     int index = 0;
