+++ Box/FullBox.h
@@ -13,10 +13,10 @@ public:
     uint32_t flags;     //lower 24 bits valid
 
 public:
-    FullBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    FullBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~FullBox();
 
-    virtual void Parse(struct Parser *parser, uint32_t start_pos) override;
+    virtual void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
