+++ Box/MediaHandleRefBox.cpp
@@ -4,13 +4,13 @@
 
 #include "MediaHandleRefBox.h"
 
-MediaHandleRefBox::MediaHandleRefBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size, type, start_pos) {}
+MediaHandleRefBox::MediaHandleRefBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size, type, start_pos) {}
 
 MediaHandleRefBox::~MediaHandleRefBox() {
 
 }
 
-void MediaHandleRefBox::Parse(struct Parser *parser, uint32_t start_pos) {
+void MediaHandleRefBox::Parse(struct Parser *parser, uint64_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     Stream *s = nullptr;
