+++ Box/MovieHeaderBox.h
@@ -43,9 +43,9 @@ public:
     uint32_t next_track_id;
 
 public:
-    MovieHeaderBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    MovieHeaderBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~MovieHeaderBox();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
