+++ Box/SampleSizeBox.cpp
@@ -4,13 +4,13 @@
 
 #include "SampleSizeBox.h"
 
-SampleSizeBox::SampleSizeBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size, type, start_pos) {}
+SampleSizeBox::SampleSizeBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size, type, start_pos) {}
 
 SampleSizeBox::~SampleSizeBox() {
 
 }
 
-void SampleSizeBox::Parse(struct Parser *parser, uint32_t start_pos) {
+void SampleSizeBox::Parse(struct Parser *parser, uint64_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     sample_size = parser->file_reader->Read32();
