+++ Box/SampleSizeBox.h
@@ -14,9 +14,9 @@ public:
     std::vector<uint32_t> sample_size_item_table;
 
 public:
-    SampleSizeBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    SampleSizeBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~SampleSizeBox();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
