+++ Box/VideoMediaHeaderBox.h
@@ -15,9 +15,9 @@ public:
     uint16_t blue;
 
 public:
-    VideoMediaHeaderBox(uint32_t size, uint32_t type, uint32_t start_pos);
+    VideoMediaHeaderBox(uint32_t size, uint32_t type, uint64_t start_pos);
     virtual ~VideoMediaHeaderBox();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
