+++ Box/VisualSampleEntry.h
@@ -11,7 +11,7 @@ class VisualSampleEntry : public Box{
 public:
     VisualSampleEntry(uint32_t size, uint32_t type, uint32_t startPos);
     virtual ~VisualSampleEntry();
-    void Parse(struct Parser *parser, uint32_t start_pos) override;
+    void Parse(struct Parser *parser, uint64_t start_pos) override;
 };
